svm-transaction: implement SVM traits for transaction views
